moved SpecialVTagKind to feat_ssr, improved its description
