+++ packages/yew/src/server_renderer.rs
@@ -9,27 +9,33 @@ use crate::html::{BaseComponent, Scope};
 use crate::platform::fmt::BufStream;
 use crate::platform::{LocalHandle, Runtime};
 
-/// Passed top-down as context for `render_into_stream` functions to know the current innermost
-/// VTag kind to apply appropriate text escaping.
 #[cfg(feature = "ssr")]
-#[derive(Default, Clone, Copy)]
-pub(crate) enum SpecialVTagKind {
-    Style,  // <style> tag
-    Script, // <script> tag
-    #[default]
-    Other,
-}
+pub(crate) mod feat_ssr {
+    /// Passed top-down as context for `render_into_stream` functions to know the current innermost
+    /// `VTag` kind to apply appropriate text escaping.
+    /// Right now this is used to make `VText` nodes aware of their environment and correctly
+    /// escape their contents when rendering them during SSR.
+    #[derive(Default, Clone, Copy)]
+    pub(crate) enum SpecialVTagKind {
+        /// <style> tag
+        Style,
+        /// <script> tag
+        Script,
+        #[default]
+        /// any other tag
+        Other,
+    }
 
-#[cfg(feature = "ssr")]
-impl<T: AsRef<str>> From<T> for SpecialVTagKind {
-    fn from(value: T) -> Self {
-        let value = value.as_ref();
-        if value.eq_ignore_ascii_case("style") {
-            Self::Style
-        } else if value.eq_ignore_ascii_case("script") {
-            Self::Script
-        } else {
-            Self::Other
+    impl<T: AsRef<str>> From<T> for SpecialVTagKind {
+        fn from(value: T) -> Self {
+            let value = value.as_ref();
+            if value.eq_ignore_ascii_case("style") {
+                Self::Style
+            } else if value.eq_ignore_ascii_case("script") {
+                Self::Script
+            } else {
+                Self::Other
+            }
         }
     }
 }
