+++ packages/yew/src/virtual_dom/vsuspense.rs
@@ -27,10 +27,10 @@ impl VSuspense {
 #[cfg(feature = "ssr")]
 mod feat_ssr {
     use super::*;
+    use crate::feat_ssr::SpecialVTagKind;
     use crate::html::AnyScope;
     use crate::platform::fmt::BufWriter;
     use crate::virtual_dom::Collectable;
-    use crate::SpecialVTagKind;
 
     impl VSuspense {
         pub(crate) async fn render_into_stream(
