+++ packages/yew/src/virtual_dom/vtext.rs
@@ -44,9 +44,9 @@ mod feat_ssr {
     use std::fmt::Write;
 
     use super::*;
+    use crate::feat_ssr::SpecialVTagKind;
     use crate::html::AnyScope;
     use crate::platform::fmt::BufWriter;
-    use crate::SpecialVTagKind;
 
     impl VText {
         pub(crate) async fn render_into_stream(
